添加单位转换 加速度单位为g 角速度单位为deg/s 加速度量程为+-12g 角速度量程为+-2000deg/s

+++ Core/Src/main.c
@@ -45,8 +45,8 @@
 /* Private variables ---------------------------------------------------------*/
 
 /* USER CODE BEGIN PV */
-int16_t rx, ry, rz;
-int16_t ax, ay, az;
+float rx, ry, rz;
+float ax, ay, az;
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
